+++ test/sentinel/test_helper.rb
@@ -1,13 +1,9 @@
 # frozen_string_literal: true
 
-$LOAD_PATH.unshift File.expand_path("../../lib", __dir__)
-require "redis-client"
-require "toxiproxy"
-
-Dir[File.join(__dir__, "../support/**/*.rb")].sort.each { |f| require f }
-Dir[File.join(__dir__, "../shared/**/*.rb")].sort.each { |f| require f }
+require_relative "../env"
 
 Servers.build_redis
 Servers::SENTINEL_TESTS.prepare
+
 require "minitest/autorun"
 Minitest.after_run { Servers::TESTS.shutdown }